# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nova — AI&U Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  prototype: 'prototype',
  in_development: 'in_development',
  deployed: 'deployed',
} as const;

export interface Project {
  id: number;
  name: string;
  slug: string;
  category: string;
  description: string;
  oneLiner: string;
  capabilities: string[];
  status: ProjectStatus;
  /** @nullable */
  technicalNote?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  whitepaper?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  labVisible: boolean;
  storefrontVisible: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  prototype: 'prototype',
  in_development: 'in_development',
  deployed: 'deployed',
} as const;

export interface ProjectInput {
  name: string;
  slug: string;
  category: string;
  description: string;
  oneLiner: string;
  capabilities?: string[];
  status: ProjectInputStatus;
  technicalNote?: string;
  linkedinUrl?: string;
  whitepaper?: string;
  price?: string;
  imageUrl?: string;
  featured?: boolean;
  labVisible?: boolean;
  storefrontVisible?: boolean;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  prototype: 'prototype',
  in_development: 'in_development',
  deployed: 'deployed',
} as const;

export interface ProjectUpdate {
  name?: string;
  category?: string;
  description?: string;
  oneLiner?: string;
  capabilities?: string[];
  status?: ProjectUpdateStatus;
  technicalNote?: string;
  linkedinUrl?: string;
  whitepaper?: string;
  price?: string;
  imageUrl?: string;
  featured?: boolean;
  labVisible?: boolean;
  storefrontVisible?: boolean;
}

export type ProjectsSummaryByCategoryItem = {
  category: string;
  count: number;
};

export type ProjectsSummaryByStatusItem = {
  status: string;
  count: number;
};

export interface ProjectsSummary {
  total: number;
  byCategory: ProjectsSummaryByCategoryItem[];
  byStatus: ProjectsSummaryByStatusItem[];
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  interest?: string | null;
  /** @nullable */
  note?: string | null;
  source: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
}

export type LeadInputSource = typeof LeadInputSource[keyof typeof LeadInputSource];


export const LeadInputSource = {
  briefing: 'briefing',
  consultation: 'consultation',
  'buy-now': 'buy-now',
  'blind-box': 'blind-box',
  'custom-request': 'custom-request',
  'hire-me': 'hire-me',
  'nova-chat': 'nova-chat',
  lab: 'lab',
} as const;

export interface LeadInput {
  name: string;
  email: string;
  organization?: string;
  role?: string;
  interest?: string;
  note?: string;
  source: LeadInputSource;
  projectId?: number;
}

export type LeadsStatsBySourceItem = {
  source: string;
  count: number;
};

export interface LeadsStats {
  total: number;
  bySource: LeadsStatsBySourceItem[];
}

export type ConversationDoor = typeof ConversationDoor[keyof typeof ConversationDoor];


export const ConversationDoor = {
  partner: 'partner',
  lab: 'lab',
  storefront: 'storefront',
} as const;

export interface Conversation {
  id: number;
  door: ConversationDoor;
  /** @nullable */
  title?: string | null;
  createdAt: string;
}

export type ConversationInputDoor = typeof ConversationInputDoor[keyof typeof ConversationInputDoor];


export const ConversationInputDoor = {
  partner: 'partner',
  lab: 'lab',
  storefront: 'storefront',
} as const;

export interface ConversationInput {
  door: ConversationInputDoor;
  title?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export type ConversationWithMessagesDoor = typeof ConversationWithMessagesDoor[keyof typeof ConversationWithMessagesDoor];


export const ConversationWithMessagesDoor = {
  partner: 'partner',
  lab: 'lab',
  storefront: 'storefront',
} as const;

export interface ConversationWithMessages {
  id: number;
  door: ConversationWithMessagesDoor;
  /** @nullable */
  title?: string | null;
  createdAt: string;
  messages: Message[];
}

export interface Decision {
  id: number;
  context: string;
  options: string[];
  chosen: string;
  prediction: string;
  confidence: number;
  primaryRisk: string;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  lesson?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DecisionInput {
  context: string;
  options: string[];
  chosen: string;
  prediction: string;
  confidence: number;
  primaryRisk: string;
}

export interface DecisionUpdate {
  outcome?: string;
  lesson?: string;
}

export type TodoStatus = typeof TodoStatus[keyof typeof TodoStatus];


export const TodoStatus = {
  approval_needed: 'approval_needed',
  waiting: 'waiting',
  completed: 'completed',
} as const;

export interface Todo {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TodoStatus;
  approved: boolean;
  createdAt: string;
  updatedAt: string;
}

export type TodoInputStatus = typeof TodoInputStatus[keyof typeof TodoInputStatus];


export const TodoInputStatus = {
  approval_needed: 'approval_needed',
  waiting: 'waiting',
  completed: 'completed',
} as const;

export interface TodoInput {
  title: string;
  description?: string;
  status: TodoInputStatus;
}

export type TodoUpdateStatus = typeof TodoUpdateStatus[keyof typeof TodoUpdateStatus];


export const TodoUpdateStatus = {
  approval_needed: 'approval_needed',
  waiting: 'waiting',
  completed: 'completed',
} as const;

export interface TodoUpdate {
  title?: string;
  description?: string;
  status?: TodoUpdateStatus;
  approved?: boolean;
}

export interface Upload {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  createdAt: string;
}

export type ListProjectsParams = {
category?: string;
status?: string;
};